+++ components/common/index.ts
@@ -1,6 +1,7 @@
 export { default as AddFriendModal } from "./AddFriendModal";
 export { default as BasicHeader } from "./BasicHeader";
 export { default as DatePicker } from "./DatePicker";
+export { default as EmptyState } from "./EmptyState";
 export { default as FriendProfileCircle } from "./FriendProfileCircle";
 export { default as GoBackBtn } from "./GoBackBtn";
 export { default as Modal } from "./Modal";
